+++ azure-storage-file/azure/storage/file/_constants.py
@@ -8,4 +8,4 @@ __author__ = 'Microsoft Corp. <ptvshelp@microsoft.com>'
 __version__ = '2.0.1'
 
 # x-ms-version for storage service.
-X_MS_VERSION = '2018-11-09'
+X_MS_VERSION = '2019-02-02'